requirements.txt
# tkinter==8.6
tk-tools==0.16.0
Pillow==7.0.0
numpy==1.24.3
tensorflow==2.13.1
keras==2.13.1
mss==9.0.1

